# docs/swagger.yaml
basePath: /
definitions:
  helper.ApiError:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  user.RegisterUserInput:
    properties:
      age:
        type: integer
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - age
    - email
    - password
    - username
    type: object
  user.UserFormatter:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  user.UserLoginFormatter:
    properties:
      token:
        type: string
    type: object
  user.UserUpdateFormatter:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: faqihyugos@gmail.com
    name: Faqih Yugos
    url: http://www.swagger.io/support
  description: Final Project FGA Golang MyGram, Go Programming Language
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Mygram Golang
  version: "1.0"
paths:
  /users:
    delete:
      consumes:
      - application/json
      description: Delete User
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ApiError'
      summary: Delete User
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update User
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserUpdateFormatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ApiError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helper.ApiError'
      summary: Update User
      tags:
      - Users
  /users/login:
    post:
      consumes:
      - application/json
      description: Login User
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserLoginFormatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ApiError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helper.ApiError'
      summary: Login User
      tags:
      - Users
  /users/register:
    post:
      consumes:
      - application/json
      description: Register User
      parameters:
      - description: User Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.RegisterUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.UserFormatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ApiError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helper.ApiError'
      summary: Register User
      tags:
      - users
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"